third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/script-defer-xhtml-expected.txt

This is a testharness.js-based test.
[FAIL] the defer script run later
  assert_equals: the script run status expected "deferred" but got "inline"
Harness: the test ran to completion.
